@@ -1,37 +1,17 @@
 module.exports = {
-    env: {
-      browser: false,
-      es6: true,
-      jest: true,
+    "env": {
+        "browser": true,
+        "es6": true
     },
-    extends: [
-      'airbnb-base',
-      'plugin:jest/all',
-    ],
-    globals: {
-      Atomics: 'readonly',
-      SharedArrayBuffer: 'readonly',
+    "extends": "eslint:recommended",
+    "globals": {
+        "Atomics": "readonly",
+        "SharedArrayBuffer": "readonly"
     },
-    parserOptions: {
-      ecmaVersion: 2018,
-      sourceType: 'module',
+    "parserOptions": {
+        "ecmaVersion": 2018,
+        "sourceType": "module"
     },
-    plugins: ['jest'],
-    rules: {
-      'max-classes-per-file': 'off',
-      'no-underscore-dangle': 'off',
-      'no-console': 'off',
-      'no-shadow': 'off',
-      'no-restricted-syntax': [
-        'error',
-        'LabeledStatement',
-        'WithStatement',
-      ],
-    },
-    overrides:[
-      {
-        files: ['*.js'],
-        excludedFiles: 'babel.config.js',
-      }
-    ]
+    "rules": {
+    }
 };
